requirements.txt
python-telegram-bot==13.12
requests
emoji
python-coinmarketcap
boto3